+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+ - Fixed Cisco conference control supported features to be accurate
+ - Implemented camera mute correctly in Cisco conference control
 
 ## [17.2.0] - 2021-01-14
 ### Added